Move the new rule to its own file.
So it (and therefore its deps) don't need to be registered in tests
that don't otherwise care about it.
