+++ src/python/pants/backend/python/rules/pex.py
@@ -1,16 +1,15 @@
 # Copyright 2019 Pants project contributors (see CONTRIBUTORS.md).
 # Licensed under the Apache License, Version 2.0 (see LICENSE).
 
-from dataclasses import dataclass
 from typing import Iterable, List, Optional, Tuple
+
+from dataclasses import dataclass
 
 from pants.backend.python.rules.download_pex_bin import DownloadedPexBin
 from pants.backend.python.rules.hermetic_pex import HermeticPex
-from pants.backend.python.rules.inject_init import InjectedInitDigest
 from pants.backend.python.subsystems.python_native_code import PexBuildEnvironment
 from pants.backend.python.subsystems.python_setup import PythonSetup
 from pants.backend.python.subsystems.subprocess_environment import SubprocessEncodingEnvironment
-from pants.engine.addressable import BuildFileAddresses
 from pants.engine.fs import (
   EMPTY_DIRECTORY_DIGEST,
   Digest,
@@ -18,12 +17,10 @@ from pants.engine.fs import (
   DirectoryWithPrefixToAdd,
 )
 from pants.engine.isolated_process import ExecuteProcessResult, MultiPlatformExecuteProcessRequest
-from pants.engine.legacy.graph import HydratedTarget, TransitiveHydratedTargets
 from pants.engine.legacy.structs import PythonTargetAdaptor, TargetAdaptor
 from pants.engine.platform import Platform, PlatformConstraint
 from pants.engine.rules import optionable_rule, rule
-from pants.engine.selectors import Get, MultiGet
-from pants.rules.core.strip_source_root import SourceRootStrippedSources
+from pants.engine.selectors import Get
 
 
 @dataclass(frozen=True)
@@ -77,14 +74,6 @@ class CreatePex:
   interpreter_constraints: PexInterpreterConstraints = PexInterpreterConstraints()
   entry_point: Optional[str] = None
   input_files_digest: Optional[Digest] = None
-
-
-@dataclass(frozen=True)
-class CreatePexFromTargetClosure:
-  """Represents a request to create a PEX from the closure of a set of targets."""
-  build_file_addresses: BuildFileAddresses
-  output_filename: str
-  entry_point: Optional[str] = None
 
 
 @dataclass(frozen=True)
@@ -153,48 +142,8 @@ async def create_pex(
   return Pex(directory_digest=result.output_directory_digest, output_filename=request.output_filename)
 
 
-@rule(name="Create PEX from targets")
-async def create_pex_from_target_closure(request: CreatePexFromTargetClosure,
-                                         python_setup: PythonSetup) -> Pex:
-  transitive_hydrated_targets = await Get[TransitiveHydratedTargets](BuildFileAddresses,
-                                                                     request.build_file_addresses)
-  all_targets = transitive_hydrated_targets.closure
-  all_target_adaptors = [t.adaptor for t in all_targets]
-
-  interpreter_constraints = PexInterpreterConstraints.create_from_adaptors(
-    adaptors=tuple(all_targets),
-    python_setup=python_setup
-  )
-
-  source_root_stripped_sources = await MultiGet(
-    Get[SourceRootStrippedSources](HydratedTarget, target_adaptor)
-    for target_adaptor in all_targets
-  )
-
-  stripped_sources_digests = [stripped_sources.snapshot.directory_digest
-                              for stripped_sources in source_root_stripped_sources]
-  sources_digest = await Get[Digest](DirectoriesToMerge(directories=tuple(stripped_sources_digests)))
-  inits_digest = await Get[InjectedInitDigest](Digest, sources_digest)
-  all_input_digests = [sources_digest, inits_digest.directory_digest]
-  merged_input_files = await Get[Digest](DirectoriesToMerge,
-                                         DirectoriesToMerge(directories=tuple(all_input_digests)))
-  requirements = PexRequirements.create_from_adaptors(all_target_adaptors)
-
-  create_pex_request = CreatePex(
-    output_filename=request.output_filename,
-    requirements=requirements,
-    interpreter_constraints=interpreter_constraints,
-    entry_point=request.entry_point,
-    input_files_digest=merged_input_files,
-  )
-
-  pex = await Get[Pex](CreatePex, create_pex_request)
-  return pex
-
-
 def rules():
   return [
     create_pex,
-    create_pex_from_target_closure,
     optionable_rule(PythonSetup),
   ]
